# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord selfbot controller API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityInputType } from './activityInputType';

export interface ActivityInput {
  type: ActivityInputType;
  /**
     * Song/track title (for Spotify) or activity name
     * @nullable
     */
  songTitle?: string | null;
  /**
     * Artist name (Spotify only)
     * @nullable
     */
  artist?: string | null;
  /**
     * Album name (Spotify only)
     * @nullable
     */
  album?: string | null;
  /**
     * Album art URL or large image URL
     * @nullable
     */
  imageUrl?: string | null;
  /**
     * Twitch channel ID or username used for streaming activity URL
     * @nullable
     */
  twitchId?: string | null;
}